agricultural_association_rules/requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Association Rules Mining
mlxtend>=0.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
bokeh>=2.4.0

# Geospatial Analysis
geopandas>=0.11.0
folium>=0.12.0
rasterio>=1.3.0
earthpy>=0.9.0

# Machine Learning Enhancement
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0

# Statistical Analysis
statsmodels>=0.13.0
pingouin>=0.5.0

# Data Processing
openpyxl>=3.0.0
xlrd>=2.0.0

# API and Data Collection
requests>=2.28.0
beautifulsoup4>=4.11.0
pyowm>=3.3.0  # OpenWeatherMap API

# Jupyter and Development
jupyter>=1.0.0
jupyterlab>=3.4.0
ipywidgets>=7.7.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=0.20.0

# Progress Bars and Utilities
tqdm>=4.64.0
rich>=12.5.0